Restructure Refer integration tests
The Refer file was getting unwieldy: hard to cross-reference existing
tests and hard to work out where to put new ones. This splits the tests
into multiple more narrowly-scoped test files to make things a bit
easier. This should also reduce the amount of rebasing required when
multiple people are working on Refer functionality

This may introduce a separate "where does this belong" issue,
but I think that's easier to live and slows us down much less than a
horrible editing and sometimes substantial rebasing experience

Approved Premises seem to do similar
