+++ integration_tests/pages/refer/index.ts
@@ -1,9 +1,11 @@
 import CheckAnswersPage from './checkAnswers'
+import CompletePage from './complete'
 import ConfirmOasysPage from './confirmOasys'
 import ConfirmPersonPage from './confirmPerson'
 import FindPersonPage from './findPerson'
 import ProgrammeHistoryPage from './programmeHistory'
 import ProgrammeHistoryDetailsPage from './programmeHistoryDetails'
+import ReasonPage from './reason'
 import SelectProgrammePage from './selectProgramme'
 import ShowPersonPage from './showPerson'
 import StartReferralPage from './startReferral'
@@ -11,11 +13,13 @@ import TaskListPage from './taskList'
 
 export {
   CheckAnswersPage,
+  CompletePage,
   ConfirmOasysPage,
   ConfirmPersonPage,
   FindPersonPage,
   ProgrammeHistoryDetailsPage,
   ProgrammeHistoryPage,
+  ReasonPage,
   SelectProgrammePage,
   ShowPersonPage,
   StartReferralPage,
